+++ src/renderer/viz/expressions/basic/constants.js
@@ -1,0 +1,7 @@
+import Constant from './constant';
+
+export const HOLD = new Constant(Number.MAX_SAFE_INTEGER);
+export const TRUE = new Constant(1);
+export const FALSE = new Constant(0);
+export const PI = new Constant(Math.PI);
+export const E = new Constant(Math.E);