+++ src/renderer/viz/expressions/placements.js
@@ -1,0 +1,5 @@
+import Placement from './placement';
+import Constant from './basic/constant';
+
+export const ALIGN_CENTER = new Placement(new Constant(0), new Constant(0));
+export const ALIGN_BOTTOM = new Placement(new Constant(0), new Constant(1));